requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
langchain
langchain-openai
langchain-qdrant
langchain-community
langgraph
qdrant-client
pypdf
openai
tiktoken
python-dotenv
loguru
pytest
httpx
pdf2image
Pillow
json-repair
